requirements.txt
# Dependências para IA conectada ao PostgreSQL
psycopg2-binary==2.9.9
openai==1.52.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-postgres==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.25
streamlit==1.29.0
plotly==5.18.0
asyncio==3.4.3
aiofiles==23.2.1
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2